- Refactor ticket validation and calculation logic to utilities and helper classes; - add unit tests for new utilities and update TicketServiceImpl to use them - refactor the maximum ticket limit from constant to properties file.
