+++ cinema-tickets-java/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+//maximum number of tickets that can be purchased at once
+max.tickets=25
